+++ R/plot_ridge.R
@@ -28,6 +28,7 @@
 #' @param theme_param list of additional params passed to `ggplot2::theme()`
 #' @param title character. Title for plot
 #' @param verbose be verbose
+#' @inheritParams gmulti_params
 #' @returns ggplot
 #' @examples
 #' g <- GiottoData::loadGiottoMini("visium")
@@ -76,7 +77,12 @@ ridgePlot <- function(gobject, feats, ...,
     save_plot = NULL,
     save_param = list(),
     default_save_name = "ridgePlot",
-    verbose = NULL) {
+    verbose = NULL,
+    view = NULL,
+    space = NULL) {
+    gobject <- .gg_materialize(gobject, view, space,
+        slots = c("cell_metadata", "expression",
+            "spatial_enrichment", "dimension_reduction"))
     value <- NULL # NSE vars
     package_check("ggridges")
     checkmate::assert_character(feats)